moved bounds code to useBoundsUpdater
